requirements.txt
selenium==4.17.2
slack_sdk==3.26.2
python-decouple==3.8
